backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3

Django==5.2.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0

dj-database-url==3.1.0
python-dotenv==1.2.1
whitenoise==6.11.0
gunicorn==23.0.0
psycopg2-binary==2.9.11

stripe==14.2.0
requests==2.32.5
urllib3==2.6.3
idna==3.11

PyJWT==2.10.1
typing_extensions==4.15.0
sqlparse==0.5.5

openai==1.12.0
